+++ src/main/java/woowacrew/article/dto/ArticleUpdateDto.java
@@ -1,4 +1,4 @@
-package woowacrew.article.domain;
+package woowacrew.article.dto;
 
 public class ArticleUpdateDto {
     private Long articleId;
