+++ src/main/java/woowacrew/article/service/ArticleInternalService.java
@@ -3,8 +3,12 @@ package woowacrew.article.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import woowacrew.article.domain.*;
+import woowacrew.article.domain.Article;
+import woowacrew.article.domain.ArticleRepository;
+import woowacrew.article.dto.ArticleRequestDto;
+import woowacrew.article.dto.ArticleUpdateDto;
 import woowacrew.article.exception.NotFoundArticleException;
+import woowacrew.article.utils.ArticleConverter;
 import woowacrew.user.domain.User;
 import woowacrew.user.dto.UserContext;
 import woowacrew.user.service.UserInternalService;
