+++ src/main/java/woowacrew/article/service/ArticleService.java
@@ -1,7 +1,11 @@
 package woowacrew.article.service;
 
 import org.springframework.stereotype.Service;
-import woowacrew.article.domain.*;
+import woowacrew.article.domain.Article;
+import woowacrew.article.dto.ArticleRequestDto;
+import woowacrew.article.dto.ArticleResponseDto;
+import woowacrew.article.dto.ArticleUpdateDto;
+import woowacrew.article.utils.ArticleConverter;
 import woowacrew.user.dto.UserContext;
 
 import java.util.List;
