+++ src/test/java/woowacrew/article/service/ArticleInternalServiceTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import woowacrew.article.domain.Article;
 import woowacrew.article.domain.ArticleRepository;
-import woowacrew.article.domain.ArticleRequestDto;
-import woowacrew.article.domain.ArticleUpdateDto;
+import woowacrew.article.dto.ArticleRequestDto;
+import woowacrew.article.dto.ArticleUpdateDto;
 import woowacrew.article.exception.MisMatchUserException;
 import woowacrew.article.exception.NotFoundArticleException;
 import woowacrew.common.service.FieldSetter;
